fix(core): keep catalog entries out of the root barrel bundle
Main gained createThemeCatalog on the ./catalog subpath (#563) after this
branch was authored, and the root core barrel re-exports that module. Rebasing
made the subpath barrel also re-export the slim entries data (#497), which
pulled the embedded catalog JSON into every root-barrel consumer and pushed
the theme-selector bundle to ~100.9KB, past its 96KB budget.

Split the curation API into catalog/create.ts; the ./catalog subpath barrel
re-exports create.js + entries.js (public API unchanged), while the root
barrel and sets.ts import create.js directly so the entries payload stays out
of bundles that don't import the ./catalog subpath. Selector bundle returns
to main parity (94,992 bytes).
