+++ packages/core/src/index.ts
@@ -11,5 +11,8 @@ export * from './themes/types.js';
 export * from './themes/metadata.js';
 export type { ThemeId } from './themes/theme-ids.js';
 export { THEME_IDS } from './themes/theme-ids.js';
-export * from './catalog/index.js';
+// Root barrel deliberately re-exports create.js, not the ./catalog barrel:
+// the precomputed entries data must stay out of bundles that only use tokens
+// (theme-selector bundle budget, #497).
+export * from './catalog/create.js';
 export * from './catalog/sets.js';
